restructure my pproject
