+++ src/components/DashboardContainer/index.ts
@@ -1,0 +1,2 @@
+import DashboardContainer from "./DashboardContainer";
+export default DashboardContainer;
