+++ src/components/ItemContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as ItemContainer } from "./ItemContainer";
