--- src/index.ts
@@ -1,9 +1,0 @@
-// Define a function to add two numbers
-export function addNumbers(a: number, b: number): number {
-  return a + b;
-}
-
-// Define a function to greet a user
-export function greetUser(name: string): string {
-  return `Hello, ${name}!`;
-}
